Fixes Linux compilation warning about unused setupFavLink function
