+++ src/common/utility.cpp
@@ -39,33 +39,7 @@
 #include <math.h>
 #include <stdarg.h>
 
-#if defined(Q_OS_WIN)
-#include "utility_win.cpp"
-#elif defined(Q_OS_MAC)
-#include "utility_mac.cpp"
-#else
-#include "utility_linux.cpp"
-#endif
-
 namespace KDC {
-
-bool OldUtility::hasSystemLaunchOnStartup(const QString &appName, log4cplus::Logger logger) {
-#if defined(Q_OS_WIN)
-    return hasSystemLaunchOnStartup_private(appName, logger);
-#else
-    Q_UNUSED(appName)
-    Q_UNUSED(logger)
-    return false;
-#endif
-}
-
-bool OldUtility::hasLaunchOnStartup(const QString &appName, log4cplus::Logger logger) {
-    return KDC::hasLaunchOnStartup_private(appName, logger);
-}
-
-void OldUtility::setLaunchOnStartup(const QString &appName, const QString &guiName, bool enable, log4cplus::Logger logger) {
-    KDC::setLaunchOnStartup_private(appName, guiName, enable, logger);
-}
 
 #ifdef Q_OS_WIN
 void OldUtility::setFolderPinState(const QUuid &clsid, bool show) {
